Correct regexes and provided scopes/language
For the single regexes of the patterns we have to select already
existing names in the theme definitions, and then append
the name of our own language. So, 'reference.code.noweb' doesn't work out of
the box because 'reference.code' doesn't exist in the already
provided themes (e.g. "dark").

In addition, the  'contributes.languages.id' property of 'package.json'
needs to match the name property in the 'tmLanguage.json' file. Also, the
'contributes.grammars.scopeName' of 'package.json' needs to match
the 'scopeName' of the 'tmLanguage.json' file.